Simplify luma downsampling for wienerns cross_filt

+++ av1/encoder/pickrst.c
@@ -4424,7 +4424,7 @@ void av1_pick_filter_restoration(const YV12_BUFFER_CONFIG *src, AV1_COMP *cpi) {
 #if CONFIG_WIENER_NONSEP
 #if CONFIG_WIENER_NONSEP_CROSS_FILT
   uint8_t *luma = NULL;
-  uint8_t *luma_buf;
+  uint16_t *luma_buf;
   const YV12_BUFFER_CONFIG *dgd = &cpi->common.cur_frame->buf;
   rsc.luma_stride = dgd->crop_widths[1] + 2 * WIENERNS_UV_BRD;
   luma_buf = wienerns_copy_luma_highbd(
